requirements.txt
aiogram
pydantic
sqlalchemy
python-dotenv
fastapi
uvicorn